Backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
SQLAlchemy==2.0.40
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.0
blinker==1.9.0
typing-extensions==4.13.2
greenlet==3.2.2
psycopg2-binary==2.9.9 